+++ zimfw.zsh
@@ -25,7 +25,7 @@
 # SOFTWARE.
 
 autoload -Uz is-at-least && if ! is-at-least 5.2; then
-  print -u2 -R $'\E[31m'${0}$': Error starting zimfw. You\'re using Zsh version \E[1m'${ZSH_VERSION}$'\E[0;31m and versions < \E[1m5.2\E[0;31m are not supported. Upgrade your Zsh.\E[0m'
+  print -u2 -R $'\E[31m'${0}$': Error starting zimfw. You\'re using Zsh version \E[1m'${ZSH_VERSION}$'\E[0;31m and versions < \E[1m5.2\E[0;31m are not supported. Update your Zsh.\E[0m'
   return 1
 fi
 autoload -Uz zargs
@@ -57,10 +57,13 @@ _zimfw_mv() {
 _zimfw_build_init() {
   local -r ztarget=${ZIM_HOME}/init.zsh
   # Force update of init.zsh if it's older than .zimrc
-  if [[ ${ztarget} -ot ${ZIM_CONFIG_FILE:-${ZDOTDIR:-${HOME}}/.zimrc} ]]; then
+  if [[ ${ztarget} -ot ${_zconfig} ]]; then
     command mv -f ${ztarget}{,.old} || return 1
   fi
   _zimfw_mv =(
+    print -R '# FILE AUTOMATICALLY GENERATED FROM '${_zconfig}
+    print '# EDIT THE SOURCE FILE AND THEN RUN zimfw build. DO NOT DIRECTLY EDIT THIS FILE!'
+    print
     print -R 'if [[ -e ${ZIM_CONFIG_FILE:-${ZDOTDIR:-${HOME}}/.zimrc} ]] zimfw() { source '${${(qqq)__ZIMFW_FILE}/${HOME}/\${HOME}}' "${@}" }'
     local zroot_dir zpre
     local -a zif_functions zif_cmds zroot_functions zroot_cmds
@@ -101,7 +104,7 @@ _zimfw_build_init() {
 _zimfw_build_login_init() {
   local -r ztarget=${ZIM_HOME}/login_init.zsh
   # Force update of login_init.zsh if it's older than .zimrc
-  if [[ ${ztarget} -ot ${ZIM_CONFIG_FILE:-${ZDOTDIR:-${HOME}}/.zimrc} ]]; then
+  if [[ ${ztarget} -ot ${_zconfig} ]]; then
     command mv -f ${ztarget}{,.old} || return 1
   fi
   _zimfw_mv =(
@@ -115,10 +118,9 @@ _zimfw_build() {
 
 _zimfw_source_zimrc() {
 zmodule() {
-  local -r ztarget=${ZIM_CONFIG_FILE:-${ZDOTDIR:-${HOME}}/.zimrc}
   local -r zusage=$'Usage: \E[1m'${0}$'\E[0m <url> [\E[1m-n\E[0m|\E[1m--name\E[0m <module_name>] [\E[1m-r\E[0m|\E[1m--root\E[0m <path>] [options]
 
-Add \E[1mzmodule\E[0m calls to your \E[1m'${ztarget}$'\E[0m file to define the modules to be initialized.
+Add \E[1mzmodule\E[0m calls to your \E[1m'${_zconfig}$'\E[0m file to define the modules to be initialized.
 The initialization will be done in the same order it\'s defined.
 
   <url>                      Module absolute path or repository URL. The following URL formats
@@ -158,6 +160,9 @@ Per-module-root options:
   \E[1m--if-command\E[0m <cmd_name>    Will only initialize module root if specified external command is
                              available. This is evaluated at every new terminal startup.
                              Equivalent to \E[1m--if \'(( \${+commands[\E[0m<cmd_name>\E[1m]} ))\'\E[0m.
+  \E[1m--if-ostype\E[0m <ostype>       Will only initialize module root if \E[1mOSTYPE\E[0m is equal to the given
+                             expression. This is evaluated at every new terminal startup.
+                             Equivalent to \E[1m--if \'[[ \${OSTYPE} == \E[0m<ostype>\E[1m ]]\'\E[0m.
   \E[1m--on-pull\E[0m <command>        Execute command after installing or updating the module. The com-
                              mand is executed in the module root directory.
   \E[1m-d\E[0m|\E[1m--disabled\E[0m              Don\'t initialize the module root or uninstall the module.
@@ -185,10 +190,6 @@ Per-call initialization options:
   other per-call initialization options, so only your provided values will be used. I.e. these
   values are either all automatic, or all manual in each zmodule call. To use default values
   and also provided values, use separate zmodule calls.'
-  if [[ ${${funcfiletrace[1]%:*}:A} != ${ztarget:A} ]]; then
-    print -u2 -lR $'\E[31m'${0}$': Must be called from \E[1m'${ztarget}$'\E[0m' '' ${zusage}
-    return 2
-  fi
   if (( ! # )); then
     print -u2 -lR $'\E[31mx '${funcfiletrace[1]}$': Missing zmodule url\E[0m' '' ${zusage}
     _zfailed=1
@@ -247,7 +248,7 @@ Per-call initialization options:
   # Set values from options
   while (( # > 0 )); do
     case ${1} in
-      -b|--branch|-t|--tag|-u|--use|--on-pull|--if|--if-command|-f|--fpath|-a|--autoload|-s|--source|-c|--cmd)
+      -b|--branch|-t|--tag|-u|--use|--on-pull|--if|--if-command|--if-ostype|-f|--fpath|-a|--autoload|-s|--source|-c|--cmd)
         if (( # < 2 )); then
           print -u2 -lR $'\E[31mx '${funcfiletrace[1]}$':\E[1m'${zname}$':\E[0;31m Missing argument for zmodule option \E[1m'${1}$'\E[0m' '' ${zusage}
           _zfailed=1
@@ -290,6 +291,10 @@ Per-call initialization options:
       --if-command)
         shift
         _zifs[${zroot_dir}]="(( \${+commands[${1}]} ))"
+        ;;
+      --if-ostype)
+        shift
+        _zifs[${zroot_dir}]="[[ \${OSTYPE} == ${1} ]]"
         ;;
       -f|--fpath)
         shift
@@ -357,14 +362,14 @@ Per-call initialization options:
 }
 
   {
-    local -r ztarget=${ZIM_CONFIG_FILE:-${ZDOTDIR:-${HOME}}/.zimrc} _zflags=${1}
+    local -r _zflags=${1}
     local -i _zfailed=0
-    if ! source ${ztarget} || (( _zfailed )); then
-      print -u2 -R $'\E[31mFailed to source \E[1m'${ztarget}$'\E[0m'
+    if ! source ${_zconfig} || (( _zfailed )); then
+      print -u2 -R $'\E[31mFailed to source \E[1m'${_zconfig}$'\E[0m'
       return 1
     fi
     if (( _zflags & 1 && ${#_znames} == 0 )); then
-      print -u2 -R $'\E[31mNo modules defined in \E[1m'${ztarget}$'\E[0m'
+      print -u2 -R $'\E[31mNo modules defined in \E[1m'${_zconfig}$'\E[0m'
       return 1
     fi
     # Remove all from _zfpaths, _zfunctions and _zcmds with disabled root dirs prefixes
@@ -462,7 +467,7 @@ _zimfw_compile() {
 }
 
 _zimfw_info() {
-  print -R 'zimfw version:        '${_zversion}' (built at 2024-09-16 23:27:18 UTC, previous commit is 3fe3ba2)'
+  print -R 'zimfw version:        '${_zversion}' (built at 2024-10-08 23:27:26 UTC, previous commit is 4bb6172)'
   local zparam
   for zparam in LANG ${(Mk)parameters:#LC_*} OSTYPE TERM TERM_PROGRAM TERM_PROGRAM_VERSION ZIM_HOME ZSH_VERSION; do
     print -R ${(r.22....:.)zparam}${(P)zparam}
@@ -680,15 +685,17 @@ _zimfw_tool_degit() {
   readonly ACTION=${1} DIR=${2} URL=${3} REV=${5} ONPULL=${7} TEMP=.zdegit_${sysparams[pid]}
   readonly TARBALL_TARGET=${DIR}/${TEMP}_tarball.tar.gz INFO_TARGET=${DIR}/.zdegit
   case ${ACTION} in
-    pre)
+    pre|prereinstall)
+      local premsg
+      if [[ ${ACTION} == pre ]] premsg=$' Use zmodule option \E[1m-z\E[0;31m|\E[1m--frozen\E[0;31m to disable this error or run \E[1mzimfw reinstall\E[0;31m to reinstall.'
       if [[ -e ${DIR} ]]; then
         if [[ ! -r ${INFO_TARGET} ]]; then
-          _zimfw_print_error $'Module was not installed using zimfw\'s degit. Use zmodule option \E[1m-z\E[0;31m|\E[1m--frozen\E[0;31m to disable this error.'
+          _zimfw_print_error $'Module was not installed using zimfw\'s degit.'${premsg}
           return 1
         fi
         readonly INFO=("${(@f)"$(<${INFO_TARGET})"}")
         if [[ ${URL} != ${INFO[1]} ]]; then
-          _zimfw_print_error 'The zimfw degit URL does not match. Expected '${URL}$'. Use zmodule option \E[1m-z\E[0;31m|\E[1m--frozen\E[0;31m to disable this error.'
+          _zimfw_print_error 'The zimfw degit URL does not match. Expected '${URL}.${premsg}
           return 1
         fi
       fi
@@ -750,14 +757,16 @@ _zimfw_tool_git() {
   readonly ACTION=${1} DIR=${2} URL=${3} TYPE=${4} ONPULL=${7}
   REV=${5}
   case ${ACTION} in
-    pre)
+    pre|prereinstall)
+      local premsg
+      if [[ ${ACTION} == pre ]] premsg=$' Use zmodule option \E[1m-z\E[0;31m|\E[1m--frozen\E[0;31m to disable this error or run \E[1mzimfw reinstall\E[0;31m to reinstall.'
       if [[ -e ${DIR} ]]; then
         if [[ ! -r ${DIR}/.git ]]; then
-          _zimfw_print_error $'Module was not installed using git. Use zmodule option \E[1m-z\E[0;31m|\E[1m--frozen\E[0;31m to disable this error.'
+          _zimfw_print_error 'Module was not installed using git.'${premsg}
           return 1
         fi
         if [[ ${URL} != $(command git -C ${DIR} config --get remote.origin.url) ]]; then
-          _zimfw_print_error 'The git URL does not match. Expected '${URL}.$' Use zmodule option \E[1m-z\E[0;31m|\E[1m--frozen\E[0;31m to disable this error.'
+          _zimfw_print_error 'The git URL does not match. Expected '${URL}.${premsg}
           return 1
         fi
       fi
@@ -846,7 +855,7 @@ _zimfw_tool_mkdir() {
   # This runs in a subshell
   readonly -i SUBMODULES=${6}
   readonly ACTION=${1} DIR=${2} TYPE=${4} REV=${5} ONPULL=${7}
-  if [[ ${ACTION} == pre ]] return 0
+  if [[ ${ACTION} == (pre|reinstall) ]] return 0
   if [[ -n ${REV} ]]; then
     _zimfw_print_warn $'The zmodule option \E[1m-'${TYPE[1]}$'\E[0;33m|\E[1m--'${TYPE}$'\E[0;33m has no effect when using the mkdir tool'
   fi
@@ -859,7 +868,8 @@ _zimfw_tool_mkdir() {
 }
 
 _zimfw_run_tool() {
-  local -r _zname=${1}
+  local zaction=${1}
+  local -r _zname=${2}
   if [[ -z ${_zurls[${_zname}]} ]]; then
     _zimfw_print_okay 'Skipping external module' 1
     return 0
@@ -874,8 +884,26 @@ _zimfw_run_tool() {
     return 1
   fi
   set "${_zdirs[${_zname}]}" "${_zurls[${_zname}]}" "${_ztypes[${_zname}]}" "${_zrevs[${_zname}]}" "${_zsubmodules[${_zname}]}" "${_zonpulls[${_zname}]}"
-  _zimfw_tool_${ztool} pre "${@}" || return 1
-  case ${_zaction} in
+  if [[ ${zaction} == reinstall ]]; then
+    _zimfw_tool_${ztool} prereinstall "${@}" && return 0
+    if (( _zprintlevel <= 0 )); then
+      command rm -rf ${_zdirs[${_zname}]} || return 1
+    else
+      local zopt
+      if (( _zprintlevel > 1 )) zopt=-v
+      if read -q "?Reinstall ${_zname} [y/N]? "; then
+        print
+        command rm -rf ${zopt} ${_zdirs[${_zname}]} || return 1
+      else
+        print
+        return 0
+      fi
+    fi
+    zaction=install
+  else
+    _zimfw_tool_${ztool} pre "${@}" || return 1
+  fi
+  case ${zaction} in
     install)
       if [[ -e ${_zdirs[${_zname}]} ]]; then
         _zimfw_print_okay 'Skipping already installed module' 1
@@ -888,30 +916,32 @@ _zimfw_run_tool() {
         _zimfw_print_error $'Not installed. Run \E[1mzimfw install\E[0;31m to install.'
         return 1
       fi
-      if [[ ${_zaction} == check ]]; then
+      if [[ ${zaction} == check ]]; then
         if (( _zprintlevel > 1 )) print -nR $'\E[2K\rChecking '${_zname}' ...'
       else
         _zimfw_print -nR $'\E[2K\rUpdating '${_zname}' ...'
       fi
       ;;
     *)
-      _zimfw_print_error "Unknown action ${_zaction}"
+      _zimfw_print_error "Unknown action ${zaction}"
       return 1
       ;;
   esac
-  _zimfw_tool_${ztool} ${_zaction} "${@}"
+  _zimfw_tool_${ztool} ${zaction} "${@}"
 }
 
 _zimfw_run_tool_action() {
-  local -r _zaction=${1}
+  local -i zmaxprocs=0
+  if [[ ${1} == reinstall ]] zmaxprocs=1
   _zimfw_source_zimrc 1 || return 1
-  zargs -n 1 -P 0 -- "${_znames[@]}" -- _zimfw_run_tool
+  zargs -n 2 -P ${zmaxprocs} -- "${_znames[@]}" -- _zimfw_run_tool ${1}
   return 0
 }
 
 zimfw() {
   builtin emulate -L zsh -o EXTENDED_GLOB
-  local -r _zversion='1.15.0-SNAPSHOT' zusage=$'Usage: \E[1m'${0}$'\E[0m <action> [\E[1m-q\E[0m|\E[1m-v\E[0m]
+  local -r _zconfig=${ZIM_CONFIG_FILE:-${ZDOTDIR:-${HOME}}/.zimrc} _zversion='1.15.0'
+  local -r zusage=$'Usage: \E[1m'${0}$'\E[0m <action> [\E[1m-q\E[0m|\E[1m-v\E[0m]
 
 Actions:
   \E[1mbuild\E[0m           Build \E[1m'${ZIM_HOME}$'/init.zsh\E[0m and \E[1m'${ZIM_HOME}$'/login_init.zsh\E[0m.
@@ -922,11 +952,14 @@ Actions:
   \E[1mcompile\E[0m         Compile Zsh files.
   \E[1mhelp\E[0m            Print this help.
   \E[1minfo\E[0m            Print zimfw and system info.
-  \E[1mlist\E[0m            List all modules currently defined in \E[1m'${ZIM_CONFIG_FILE:-${ZDOTDIR:-${HOME}}/.zimrc}$'\E[0m.
+  \E[1mlist\E[0m            List all modules currently defined in \E[1m'${_zconfig}$'\E[0m.
                   Use \E[1m-v\E[0m to also see the modules details.
   \E[1minit\E[0m            Same as \E[1minstall\E[0m, but with output tailored to be used at terminal startup.
   \E[1minstall\E[0m         Install new modules. Also does \E[1mbuild\E[0m, \E[1mcompile\E[0m. Use \E[1m-v\E[0m to also see their
                   output, any on-pull output and skipped modules.
+  \E[1mreinstall\E[0m       Reinstall modules that failed check. Prompts for confirmation. Use \E[1m-q\E[0m for
+                  quiet reinstall. Also does \E[1mbuild\E[0m, \E[1mcompile\E[0m. Use \E[1m-v\E[0m to also see their output,
+                  any on-pull output and skipped modules.
   \E[1muninstall\E[0m       Delete unused modules. Prompts for confirmation. Use \E[1m-q\E[0m for quiet uninstall.
   \E[1mcheck\E[0m           Check if updates for current modules are available. Use \E[1m-v\E[0m to also see
                   skipped and up to date modules.
@@ -966,7 +999,7 @@ Options:
     _zimfw_check_version ${zversion_check_force} 1
   fi
 
-  if [[ ! -w ${ZIM_HOME} && ${1} == (build|check|init|install|update|check-version) ]]; then
+  if [[ ! -w ${ZIM_HOME} && ${1} == (build|check|init|install|update|reinstall|check-version) ]]; then
     print -u2 -R $'\E[31m'${0}$': No write permission to \E[1m'${ZIM_HOME}$'\E[0;31m. Will not try to '${1}$'.\E[0m'
     return 1
   fi
@@ -1001,7 +1034,7 @@ Options:
       _zimfw_print 'Done with install.' # Only printed in verbose mode
       _zimfw_source_zimrc 2 && _zimfw_build && _zimfw_compile
       ;;
-    install|update)
+    install|update|reinstall)
       _zimfw_run_tool_action ${1} || return 1
       _zimfw_print -R "Done with ${1}.${_zrestartmsg}"
       (( _zprintlevel-- ))